+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - new validation methods `Essence::validateAll`, `Essence::validate` (alias of `Essence::go`) +
 - new configuration methods `Essence::addLink` and `Essence::addMatcher` +
 - wrapper `expect` now works with other data types (not only `Essence` instances) +
-- `AbstractMatcher` now will automatically replace matcher's `value` property value
+- `AbstractMatcher` now will automatically replace matcher's `value` property value +
 
 ## Changes